Tidied up a couple of comments, warnings and did a bit of code modernisation

+++ modules/juce_audio_basics/midi/juce_MidiMessage.h
@@ -73,7 +73,7 @@ public:
     MidiMessage (int byte1, int byte2, int byte3, Data... otherBytes)  : size (3 + sizeof... (otherBytes))
     {
         // this checks that the length matches the data..
-        jassert (size > 3 || byte1 >= 0xf0 || getMessageLengthFromFirstByte (byte1) == size);
+        jassert (size > 3 || byte1 >= 0xf0 || getMessageLengthFromFirstByte ((uint8) byte1) == size);
 
         const uint8 data[] = { (uint8) byte1, (uint8) byte2, (uint8) byte3, static_cast<uint8> (otherBytes)... };
         memcpy (allocateSpace (size), data, (size_t) size);
